Move adjust_indentation to a shared home
