+++ crates/ruff_linter/src/rules/pylint/rules/collapsible_else_if.rs
@@ -9,7 +9,7 @@ use ruff_source_file::Locator;
 use ruff_text_size::{Ranged, TextRange};
 
 use crate::checkers::ast::Checker;
-use crate::rules::pyupgrade::fixes::adjust_indentation;
+use crate::fix::edits::adjust_indentation;
 
 /// ## What it does
 /// Checks for `else` blocks that consist of a single `if` statement.
