+++ crates/ruff_linter/src/rules/pyupgrade/rules/outdated_version_block.rs
@@ -1,6 +1,7 @@
 use std::cmp::Ordering;
 
 use anyhow::Result;
+
 use ruff_diagnostics::{Diagnostic, Edit, Fix, FixAvailability, Violation};
 use ruff_macros::{derive_message_formats, violation};
 use ruff_python_ast::helpers::map_subscript;
@@ -10,9 +11,7 @@ use ruff_python_ast::{self as ast, CmpOp, ElifElseClause, Expr, Int, StmtIf};
 use ruff_text_size::{Ranged, TextLen, TextRange};
 
 use crate::checkers::ast::Checker;
-use crate::fix::edits::delete_stmt;
-
-use crate::rules::pyupgrade::fixes::adjust_indentation;
+use crate::fix::edits::{adjust_indentation, delete_stmt};
 use crate::settings::types::PythonVersion;
 
 /// ## What it does
